fix clear tmp directory

+++ app/UseCases/Admin/Shop/DailyImportService.php
@@ -671,7 +671,7 @@ class DailyImportService
                     }
                     app(ProductController::class)->store($request);
                 }
-                if (isset($localImg)) {
+                if (isset($localImg) && file_exists($localImg)) {
                     unlink($localImg);
                 }
                 echo 'Товар: '. $mainCategory->name . ' ' . $product['sku'] . ' сохранен в базу данных'.PHP_EOL;
